Mobile responsiveness/cosmetics patch
Changes:
- Changed anchor tag colour, to differentiate links from the text.
- Added an anchor:hover style to convey clickability.

Fixes:
- Changed background colour to better blend with background image.
- Added black fade to bottom of background image, so it blends better
with the background.
- Modified position and margin rules on .flipbook-viewport and
#kronorium, so that the book is always centered on any screen size.
- Prototyped a replacement jQuery method for .offset(), as this method
is unreliable if the viewport is zoomed. This was breaking page-flipping
functionality on mobile. This method is called .offset_FIXED() and is
implemented and documented in "turn.js.mobilefix.js".
- Modified "turn.min.js" and replaced the .offset() call at line 43 col
466 with the aforementioned .offset_FIXED(). Page turning now works
properly on mobile, regardless of the zoom/offset of the viewport.
(Renamed file "turn.min.fixed.js" as to serve as a reminder if that
library ever needs to be replaced)